Organize src 📦
