+++ src/components/Navigation/index.ts
@@ -1,0 +1,3 @@
+import Navigation from './Navigation';
+
+export default Navigation;
