+++ src/index.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom';
 
 import { AppContainer } from 'react-hot-loader';
 
-import App from './App';
+import App from './components/App';
 
 import 'semantic-ui-css/semantic.min.css';
 import './scss/index.scss';
