+++ src/lib/interfaces.d.ts
